Migrate to mercurius/fastify
